Made SearchBar its own component
