+++ client/src/components/SearchBar/SearchBar.module.css
@@ -1,0 +1,27 @@
+.Search {
+  align-items: center;
+  display: flex;
+  font-size: 1.5rem;
+  height: 3rem;
+  justify-content: center;
+  width: 100%;
+}
+
+.SearchBar {
+  border-radius: 1rem;
+  font-size: 1rem;
+  height: 100%;
+  text-align: center;
+  width: 100%;
+}
+
+.SearchButton {
+  background-color: green;
+  border-radius: 10px;
+  cursor: pointer;
+  font-size: 1.5rem;
+  height: 100%;
+  line-height: 3rem;
+  margin-left: 1rem;
+  padding: 0 0.5rem 0 0.5rem;
+}
